pdf_generator/requirements.txt
azure-functions
reportlab